+++ src/main/java/by/netcracker/artemyev/service/GenderService.java
@@ -12,6 +12,6 @@ public interface GenderService extends GeneralService<Gender> {
     void add(Gender object) throws ServiceException;
     void update(Gender object) throws ServiceException;
     void remove(Gender object) throws ServiceException;
-    Gender getById(int id) throws ServiceException;
+    Gender getById(Long id) throws ServiceException;
     List<Gender> getAll() throws ServiceException;
 }
