+++ src/main/java/by/netcracker/artemyev/service/GeneralService.java
@@ -11,6 +11,6 @@ public interface GeneralService<T> {
     void add(T object) throws ServiceException;
     void update(T object) throws ServiceException;
     void remove(T object) throws ServiceException;
-    T getById(int id) throws ServiceException;
+    T getById(Long id) throws ServiceException;
     List<T> getAll() throws ServiceException;
 }
