+++ src/main/java/by/netcracker/artemyev/service/TeamService.java
@@ -12,6 +12,6 @@ public interface TeamService extends GeneralService<Team> {
     void add(Team object) throws ServiceException;
     void update(Team object) throws ServiceException;
     void remove(Team object) throws ServiceException;
-    Team getById(int id) throws ServiceException;
+    Team getById(Long id) throws ServiceException;
     List<Team> getAll() throws ServiceException;
 }
